requirements.txt
torch
torchvision
timm
gradio
grad-cam
Pillow
numpy
matplotlib
